worker/requirements.txt
instagrapi==2.1.3
redis==5.0.7
PyMySQL==1.1.1
python-dotenv==1.0.1
langdetect==1.0.9
sentry-sdk==2.13.0
requests==2.32.3
Pillow==10.4.0
